Loop natively and use authenticated logins
This patch makes the tool run forever, looping over all the projects
every minute. In addition, it now uses authenticated access as it will
sign-in as the user of the current machine. This makes it suitable to
process merge requests from a private project. The loop makes deployment
easier though there are still unsolved issues that prevent this from
really running as a service.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>