Add experimental feedback to git merges
This patch adds experimental feedback to each merged git branch. Upon
initiation of the merge, a comment is added to let the user know that
pmr is actively looking at merge request.

Following that, preparations are made (clones/fetches) and if they fail
an error is logged (but nothing is mentioned in the merge request yet as
the issue may be temporary).

Next a merge is attempted, if it fails an appropriate comment is added
and (with the vote of 'needs fixing') and the whole proposal is reverted
from approved to needs reviewed.

If the merge works, tests are started (optionally). Again if they fail
an appropriate comment is added and the merge proposal is updated.

Finally we try to push the branch, if that fails we just fail silently
as the whole process will begin again and this might be a transient
failure.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>