Produce nicer git commit messages
This avoids ugly "Merge commit '<hash>' into <merge-id>" messages, and
causes the commit_message field on merge proposals to mean something.
We now always create a merge commit, and construct a commit message that
includes the merge proposal ID, a shortened version of the repository
and path that match the titles of merge proposals in the Launchpad web
UI, and the commit_message field from the merge proposal.  For example,
one might end up with:

  Merge #300000 from ~contributor/project:branch

  Rotate weasels instead of stoats

  Stoats object to being rotated, so rotate weasels instead.  Unless
  otherwise specified, mustelid-rotation is through 180 degrees.