Add a mutually_exclusive_option group to handle projects
Projects can still be provided on the command line with --project_list.
But for large list of projects, a configuration file is more suitable.

The syntax of such pmr.conf file is:

    [PMR]

    [lp:plainbox-provider-certification-client]
    [lp:plainbox-provider-certification-server]

Note that all projects have to be prefixed by "lp:". More than a convention
it's a way to select all section names in one go and keep the top level section
for future needs.